# services/news-feed-engine/api/openapi.yaml
openapi: 3.1.0
info:
  title: ElevatedIQ News Feed Engine API
  description: |
    AI-powered news aggregation platform API for multi-platform content ingestion,
    analysis, and video summary generation. Supports white-label multi-tenant deployment.
  version: 1.0.0
  contact:
    name: ElevatedIQ Platform Team
    email: platform@elevatediq.ai
  license:
    name: Proprietary
servers:
  - url: https://api.elevatediq.ai/news-feed/v1
    description: Production
  - url: https://staging-api.elevatediq.ai/news-feed/v1
    description: Staging
  - url: http://127.0.0.1:8096/api/v1
    description: Local Development
...
paths:
...
  /analytics/content:
    get:
      tags: [Analytics]
      summary: Get content analytics
      operationId: getContentAnalytics
      security:
        - bearerAuth: []
      parameters:
        - name: range
          in: query
          schema:
            type: string
            enum: [1d, 7d, 30d, 90d]
            default: 7d
        - $ref: '#/components/parameters/category'
        - $ref: '#/components/parameters/platform'
      responses:
        '200':
          description: Content analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentAnalyticsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /analytics/creators:
    get:
      tags: [Analytics]
      summary: Get creator analytics
      operationId: getCreatorAnalytics
      security:
        - bearerAuth: []
      parameters:
        - name: range
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 30d
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Creator analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatorAnalyticsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/config:
    get:
      tags: [Admin]
      summary: Get service configuration
      operationId: getConfig
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Service configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Admin]
      summary: Update service configuration
      operationId: updateConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/analytics/overview:
    get:
      tags: [Admin]
      summary: Get analytics overview
      operationId: getAnalyticsOverview
      security:
        - bearerAuth: []
      parameters:
        - name: range
          in: query
          schema:
            type: string
            enum: [1d, 7d, 30d, 90d]
            default: 7d
      responses:
        '200':
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsOverview'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    category:
      name: category
      in: query
      schema:
        type: string
    platform:
      name: platform
      in: query
      schema:
        type: string
        enum: [youtube, twitter, reddit, rss, internal]
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Creator:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        platform:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        tier:
          type: string
          enum: [platinum, gold, silver, bronze, unverified]
        verified_at:
          type: string
          format: date-time
        follower_count:
          type: integer
        engagement_rate:
          type: number
          format: float
...
    AnalyticsOverview:
      type: object
      properties:
        tenant_id:
          type: string
        time_range:
          type: string
        total_content:
          type: integer
        content_this_period:
          type: integer
        total_creators:
          type: integer
        total_videos:
          type: integer
        total_views:
          type: integer
        engagement_rate:
          type: number
        top_categories:
          type: array
          items:
            type: object
        platform_breakdown:
          type: array
          items:
            type: object
...
    ConfigResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string

    # AI Schemas
...
    ContentAnalyticsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            time_range:
              type: string
            total_content:
              type: integer
            content_by_category:
              type: array
              items:
                type: object
                properties:
                  category:
                    type: string
                  count:
                    type: integer
                  percentage:
                    type: number
                    format: float
            content_by_platform:
              type: array
              items:
                type: object
                properties:
                  platform:
                    type: string
                  count:
                    type: integer
            content_over_time:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  count:
                    type: integer
            avg_quality_score:
              type: number
              format: float
            avg_engagement_score:
              type: number
              format: float
...
    CreatorAnalyticsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            time_range:
              type: string
            total_creators:
              type: integer
            top_creators:
              type: array
              items:
                type: object
                properties:
                  creator:
                    $ref: '#/components/schemas/Creator'
                  content_count:
                    type: integer
                  total_engagement:
                    type: integer
                  avg_quality_score:
                    type: number
                    format: float
            creator_growth:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  new_creators:
                    type: integer
            tier_distribution:
              type: array
              items:
                type: object
                properties:
                  tier:
                    type: string
                  count:
                    type: integer
...